fix: Distinguish between timeout and concept id error.
Previously both the DISS timeout and the to_concept function
could produce a concept id error. This lead to a race condition
where we could have a timeout when handling a concept id error
from to_concept. This patch changes the DISS timeout to use
the TimeoutError and catch this seperately.